✨: redis를 통한 검색 키워드를 캐싱하고 db를 통해 조회하는 자동완성 구현

+++ src/main/java/com/war11/domain/product/service/ProductService.java
@@ -4,6 +4,7 @@ import com.war11.domain.product.dto.request.ProductAutoCompletingRequest;
 import com.war11.domain.product.dto.request.ProductFindRequest;
 import com.war11.domain.product.dto.request.ProductRequest;
 import com.war11.domain.product.dto.request.ProductUpdateRequest;
+import com.war11.domain.product.dto.response.KeywordResponse;
 import com.war11.domain.product.dto.response.ProductResponse;
 import com.war11.domain.product.entity.Keyword;
 import com.war11.domain.product.entity.Product;
@@ -17,10 +18,7 @@ import java.util.List;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.domain.Sort.Order;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.ZSetOperations;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -35,7 +33,8 @@ public class ProductService {
   private final KeywordRepository keywordRepository;
   private final RedisTemplate<String,String> redisTemplate;
 
-  public ProductResponse createProduct(ProductRequest productRequest) {
+  public ProductResponse createProduct(ProductRequest productRequest, String username) {
+    checkAdmin(username);
 
     Product resultProduct = productRepository.save(productRequest.toEntity(productRequest));
     return resultProduct.toDto(resultProduct);
@@ -43,72 +42,51 @@ public class ProductService {
   }
 
   @Transactional
-  public ProductResponse updateProduct(ProductUpdateRequest productUpdateRequest) {
+  public ProductResponse updateProduct(ProductUpdateRequest productUpdateRequest, String username) {
+    checkAdmin(username);
 
-    productValidationById(productUpdateRequest.id());
+    findByIdOrThrow(productUpdateRequest.id());
     Product resultProduct = productRepository.findByIdForUpdate(productUpdateRequest.id());
     resultProduct.updateProduct(productUpdateRequest);
     return resultProduct.toDto(resultProduct);
   }
 
   @Transactional
-  public void deleteProduct(Long productId) {
+  public void deleteProduct(Long productId, String username) {
+    checkAdmin(username);
 
-    productValidationById(productId);
+    findByIdOrThrow(productId);
     Product resultProduct = productRepository.findByIdForUpdate(productId);
     resultProduct.deleteProduct();
   }
 
   public ProductResponse findByProductId(Long productId) {
-    Product result = productValidationById(productId);
+    Product result = findByIdOrThrow(productId);
     return result.toDto(result);
   }
 
-  private Product productValidationById(Long productId) {
+  private Product findByIdOrThrow(Long productId) {
     return productRepository.findById(productId).
         orElseThrow(() ->
             new BusinessException(ErrorCode.NOT_FOUND_PRODUCT_ID));
   }
 
   @Transactional
-  public Page<ProductResponse> findByProductName(ProductFindRequest productFindRequest) {
+  public Page<ProductResponse> findByProductName(ProductFindRequest productFindRequest,Pageable pageable) {
 
-    Order order;
-    if (productFindRequest.getOrder().equals("asc")) {
-      order = Sort.Order.asc("updatedAt");
-    } else {
-      order = Sort.Order.desc("updatedAt");
-    }
-
-    Pageable pageable = PageRequest.of(
-        productFindRequest.getPage() - 1,
-        productFindRequest.getSize(),
-        Sort.by(order));
 
     ZSetOperations<String,String> zSetOperations = redisTemplate.opsForZSet();
-    zSetOperations.incrementScore("keyword",productFindRequest.getName(),1);
-
-
-//    if (keywordRepository.existsById(productFindRequest.getName())) {
-//      Keyword keyword = keywordRepository.findById(productFindRequest.getName()).orElseThrow(()
-//          -> new BusinessException(ErrorCode.NOT_FOUND_KEYWORD_ID));
-//      keyword.incrementCount();
-//    } else {
-//      keywordRepository.save(new Keyword(productFindRequest.getName(), 1L));
-//    }
+    zSetOperations.incrementScore("keyword",productFindRequest.name(),1);
 
     return productRepository.findByProductName(productFindRequest, pageable);
   }
 
-  public Page<Keyword> findByAutoCompleting(ProductAutoCompletingRequest productAutoCompletingRequest) {
-    Order order = Sort.Order.desc("count");
+  public Page<KeywordResponse> findByAutoCompleting(ProductAutoCompletingRequest productAutoCompletingRequest,Pageable pageable) {
+    LocalDateTime end = LocalDateTime.now();
+    LocalDateTime start = end.minusHours(1);
 
-    Pageable pageable = PageRequest.of(
-        productAutoCompletingRequest.getPage() - 1,
-        productAutoCompletingRequest.getSize(),
-        Sort.by(order));
-
-    return keywordRepository.findByKeywordStartingWith(productAutoCompletingRequest.getKeyword(),pageable);
+     return keywordRepository.findByKeyword(
+        productAutoCompletingRequest.keyword(), start, end, pageable);
   }
 
   @Scheduled(fixedRate = 300000)
@@ -125,5 +103,12 @@ public class ProductService {
 
     keywordRepository.saveAll(keywordList);
     redisTemplate.delete("keyword");
+  }
+
+  //Create,Update,Delete 모두 권한 검사가 핅요하므로, 중복코드 방지를 위해 통합.
+  public void checkAdmin(String  username){
+    if(!username.equals("ADMIN")){
+      throw new BusinessException(ErrorCode.UMATHORIZED_ADMIN);
+    }
   }
 }
